removes quickActions from activity_main.xml
