+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     package="com.example.notes">
 
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
+    <uses-permission android:name="android.permission.CAMERA" />
 
     <application
         android:allowBackup="true"
